# Conflicts:
#	Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
#	Assets/Resources/GeneratedWeapons/Fast Gun.asset
#	ProjectSettings/EditorBuildSettings.asset
#	fmod_editor.log

+++ Assets/Scripts/Kiwii/PlayerData.cs
@@ -8,7 +8,7 @@ public class PlayerData
 {
     public int CurrentLvl;
 
-    public float PlayerMaxHealth;
+    public int PlayerMaxHealth;
 
     public int PlayerCurrentMoney;
